pyproject.toml
[tool.poetry]
name = "es-base-engines"
version = "0.1.0"
description = ""
authors = ["oktapian <oktapian@jkt1.ebdesk.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "3.10.12"
elasticsearch7 = "^7.17.12"
module-typica = "^0.2.2"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
